delete board, edit board, install jquery confirm plugin

+++ public/js/app.js
@@ -25,6 +25,7 @@ angular.module('app').controller('boardsCtrl', function($scope, boardsFactory, $
         var modalInstance = $uibModal.open({
             templateUrl: '/api/boards-create-form',
             controller: function($scope, $uibModalInstance) {
+                $scope.formTitle = 'Create board';
                 $scope.boardSubmit = function() {
                     if($scope.boardForm.$valid) {
                         boardsFactory.createBoard($scope.board)
@@ -46,13 +47,51 @@ angular.module('app').controller('boardsCtrl', function($scope, boardsFactory, $
         });
     };
 
+    $scope.editBoard = function(id) {
+        var modalInstance = $uibModal.open({
+            templateUrl: '/api/boards-create-form',
+            controller: function($scope, $uibModalInstance) {
+                $scope.formTitle = 'Edit board';
+                boardsFactory.editBoard(id)
+                    .then(function(rec) {
+                        var currentBoard = angular.element(document.querySelector('#board_name'));
+                        currentBoard.val(rec.name);
+                        $scope.boardSubmit = function() {
+                            if($scope.boardForm.$valid) {
+                                boardsFactory.updateBoard(id, $scope.board)
+                                    .then(function() {
+                                        $uibModalInstance.close();
+                                    });
+                            }
+                        };
+                    });
+
+                $scope.cancel = function() {
+                    $uibModalInstance.dismiss('cancel');
+                };
+            }
+        });
+
+        modalInstance.result.then(function() {
+            $scope.allBoards();
+        }, function() {
+            // error
+        });
+    };
+
     $scope.deleteBoard = function(id) {
-        boardsFactory.deleteBoard(id)
-            .then(function(rec) {
-                $scope.allBoards();
-            }, function(err) {
-                // error
-            });
+        $.confirm({
+            title: 'Are you sure?',
+            content: 'The board to move to a basket!',
+            confirm: function(){
+                boardsFactory.deleteBoard(id)
+                    .then(function(rec) {
+                        $scope.allBoards();
+                    }, function(err) {
+                        // error
+                    });
+            }
+        });
     };
 
     $scope.init();
@@ -89,10 +128,36 @@ angular.module('app').factory('boardsFactory', function($http, $q) {
         return defer.promise;
     };
 
+    board.editBoard = function(id) {
+        var defer = $q.defer();
+        $http.get('/api/boards/' + id + '/edit')
+            .success(function(rec) {
+                defer.resolve(rec);
+            })
+            .error(function(err, status) {
+                defer.reject(err);
+            });
+
+        return defer.promise;
+    };
+
+    board.updateBoard = function(id, board) {
+        var defer = $q.defer();
+        $http.put('/api/boards/' + id, board)
+            .success(function(rec) {
+                defer.resolve(rec);
+            })
+            .error(function(err, status) {
+                defer.reject(err);
+            });
+
+        return defer.promise;
+    };
+
     board.deleteBoard = function(id) {
         var defer = $q.defer();
 
-        $http.delete('/api/boards/destroy/' + id)
+        $http.delete('/api/boards/' + id)
             .success(function(rec) {
                 defer.resolve(rec);
             })
