+++ resources/assets/js/controllers/boardsCtrl.js
@@ -20,6 +20,7 @@ angular.module('app').controller('boardsCtrl', function($scope, boardsFactory, $
         var modalInstance = $uibModal.open({
             templateUrl: '/api/boards-create-form',
             controller: function($scope, $uibModalInstance) {
+                $scope.formTitle = 'Create board';
                 $scope.boardSubmit = function() {
                     if($scope.boardForm.$valid) {
                         boardsFactory.createBoard($scope.board)
@@ -41,13 +42,51 @@ angular.module('app').controller('boardsCtrl', function($scope, boardsFactory, $
         });
     };
 
+    $scope.editBoard = function(id) {
+        var modalInstance = $uibModal.open({
+            templateUrl: '/api/boards-create-form',
+            controller: function($scope, $uibModalInstance) {
+                $scope.formTitle = 'Edit board';
+                boardsFactory.editBoard(id)
+                    .then(function(rec) {
+                        var currentBoard = angular.element(document.querySelector('#board_name'));
+                        currentBoard.val(rec.name);
+                        $scope.boardSubmit = function() {
+                            if($scope.boardForm.$valid) {
+                                boardsFactory.updateBoard(id, $scope.board)
+                                    .then(function() {
+                                        $uibModalInstance.close();
+                                    });
+                            }
+                        };
+                    });
+
+                $scope.cancel = function() {
+                    $uibModalInstance.dismiss('cancel');
+                };
+            }
+        });
+
+        modalInstance.result.then(function() {
+            $scope.allBoards();
+        }, function() {
+            // error
+        });
+    };
+
     $scope.deleteBoard = function(id) {
-        boardsFactory.deleteBoard(id)
-            .then(function(rec) {
-                $scope.allBoards();
-            }, function(err) {
-                // error
-            });
+        $.confirm({
+            title: 'Are you sure?',
+            content: 'The board to move to a basket!',
+            confirm: function(){
+                boardsFactory.deleteBoard(id)
+                    .then(function(rec) {
+                        $scope.allBoards();
+                    }, function(err) {
+                        // error
+                    });
+            }
+        });
     };
 
     $scope.init();
